requirements.txt
backcall==0.1.0
colorama==0.3.9
decorator==4.3.0
Django==1.11
django-debug-toolbar==1.9.1
djangorestframework==3.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
mysqlclient==1.3.12
parso==0.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
Pygments==2.2.0
pytz==2018.4
shell==1.0.1
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
wcwidth==0.1.7
